% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monty_hall_sim.R
\name{monty_hall_sim}
\alias{monty_hall_sim}
\title{Simulate the Monty Hall Problem}
\usage{
monty_hall_sim(n_doors, n_simulations)
}
\arguments{
\item{n_doors}{Number of doors (>= 3).}

\item{n_simulations}{Number of simulations.}
}
\value{
A list with sticking and switching win rates.
}
\description{
Runs a simulation of the Monty Hall problem with a specified number of doors.
}
\examples{
monty_hall_sim(3, 100)
}
