% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_mhp_app}
\alias{run_mhp_app}
\title{Run Monty Hall Shiny App}
\usage{
run_mhp_app()
}
\value{
Launches a Shiny app.
}
\description{
Launches an interactive Shiny app for Monty Hall simulations.
}
\examples{
\donttest{
if (interactive()) run_mhp_app()
}
}
